Create visibility task in internal visibility queue (#988)

+++ service/history/timerQueueTaskExecutorBase.go
@@ -119,9 +119,13 @@ func (t *timerQueueTaskExecutorBase) executeDeleteHistoryEventTask(
 
 func (t *timerQueueTaskExecutorBase) deleteWorkflow(
 	task *persistencespb.TimerTaskInfo,
-	context workflowExecutionContext,
+	workflowContext workflowExecutionContext,
 	msBuilder mutableState,
 ) error {
+
+	if err := t.deleteWorkflowVisibility(task, workflowContext, msBuilder); err != nil {
+		return err
+	}
 
 	if err := t.deleteCurrentWorkflowExecution(task); err != nil {
 		return err
@@ -135,12 +139,9 @@ func (t *timerQueueTaskExecutorBase) deleteWorkflow(
 		return err
 	}
 
-	if err := t.deleteWorkflowVisibility(task); err != nil {
-		return err
-	}
 	// calling clear here to force accesses of mutable state to read database
 	// if this is not called then callers will get mutable state even though its been removed from database
-	context.clear()
+	workflowContext.clear()
 	return nil
 }
 
@@ -187,23 +188,26 @@ func (t *timerQueueTaskExecutorBase) archiveWorkflow(
 		return err
 	}
 
+	// delete visibility record here regardless if it's been archived inline or not
+	// since the entire record is included as part of the archive request.
+	if err := t.deleteWorkflowVisibility(task, workflowContext, msBuilder); err != nil {
+		return err
+	}
+
 	if err := t.deleteCurrentWorkflowExecution(task); err != nil {
 		return err
 	}
+
 	if err := t.deleteWorkflowExecution(task); err != nil {
 		return err
 	}
+
 	// delete workflow history if history archival is not needed or history as been archived inline
 	if resp.HistoryArchivedInline {
 		t.metricsClient.IncCounter(metrics.HistoryProcessDeleteHistoryEventScope, metrics.WorkflowCleanupDeleteHistoryInlineCount)
 		if err := t.deleteWorkflowHistory(task, msBuilder); err != nil {
 			return err
 		}
-	}
-	// delete visibility record here regardless if it's been archived inline or not
-	// since the entire record is included as part of the archive request.
-	if err := t.deleteWorkflowVisibility(task); err != nil {
-		return err
 	}
 	// calling clear here to force accesses of mutable state to read database
 	// if this is not called then callers will get mutable state even though its been removed from database
@@ -260,19 +264,39 @@ func (t *timerQueueTaskExecutorBase) deleteWorkflowHistory(
 
 func (t *timerQueueTaskExecutorBase) deleteWorkflowVisibility(
 	task *persistencespb.TimerTaskInfo,
+	workflowContext workflowExecutionContext,
+	msBuilder mutableState,
 ) error {
 
-	op := func() error {
-		request := &persistence.VisibilityDeleteWorkflowExecutionRequest{
-			NamespaceID: task.GetNamespaceId(),
-			WorkflowID:  task.GetWorkflowId(),
-			RunID:       task.GetRunId(),
-			TaskID:      task.GetTaskId(),
+	if !t.config.DisableKafkaForVisibility() {
+		op := func() error {
+			request := &persistence.VisibilityDeleteWorkflowExecutionRequest{
+				NamespaceID: task.GetNamespaceId(),
+				WorkflowID:  task.GetWorkflowId(),
+				RunID:       task.GetRunId(),
+				TaskID:      task.GetTaskId(),
+			}
+			// TODO: expose GetVisibilityManager method on shardContext interface
+			return t.shard.GetService().GetVisibilityManager().DeleteWorkflowExecution(request) // delete from db
 		}
-		// TODO: expose GetVisibilityManager method on shardContext interface
-		return t.shard.GetService().GetVisibilityManager().DeleteWorkflowExecution(request) // delete from db
+		return backoff.Retry(op, persistenceOperationRetryPolicy, common.IsPersistenceTransientError)
 	}
-	return backoff.Retry(op, persistenceOperationRetryPolicy, common.IsPersistenceTransientError)
+
+	now := t.shard.GetTimeSource().Now()
+	msBuilder.AddVisibilityTasks(&persistence.DeleteExecutionVisibilityTask{
+		// TaskID is set by shard
+		VisibilityTimestamp: now,
+		Version:             task.GetVersion(),
+	})
+
+	return workflowContext.updateWorkflowExecutionWithNew(
+		now,
+		persistence.UpdateWorkflowModeBypassCurrent,
+		nil, // no new workflow
+		nil, // no new workflow
+		transactionPolicyPassive,
+		nil,
+	)
 }
 
 func (t *timerQueueTaskExecutorBase) getNamespaceIDAndWorkflowExecution(
